+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- [TD-7011] Add i18n content non-translatable fields
+
+
 ## [6.16.0] 2024-12-16
 
 ### Added